improve: PR #42 개선 제안 — JPQL → QueryDSL + I2·I3·I5·I6 반영
리뷰어 지적 6건 중 I1(JPQL 스타일) 은 QueryDSL 전환으로 대체, I4(페이로드)는 MVP 유지.

### 핵심: JPQL → QueryDSL
SessionAttempt 의 집계 쿼리 3건을 Spring Data `@Query` JPQL 에서
`SessionAttemptRepositoryCustom` + `SessionAttemptRepositoryImpl` 의
QueryDSL 패턴으로 재작성 (기존 Gym/Route/ClimbingSession 커스텀 규약과 동일).

- cross-join 표기 (`FROM a, s WHERE a.sessionId = s.id`) → 명시적 `JOIN s ON a.sessionId = s.id`
- 공통 where 조합을 `ownedNotDeletedAndResultIn` 헬퍼로 추출
- `findTopGradeValueByUserId` 반환이 `List<String>` + `Pageable.ofSize(1)` 에서
  `Optional<String>` + `.limit(1)` 로 간결화

### I2. topGrade 동률 보조 정렬
`ORDER BY gradeNumeric DESC, loggedAt DESC` — 동일 gradeNumeric 시 최근 시도 우선.

### I3. 카운트 타입 int → long 통일
`MeStatsView.weekSessions/weekSends` 와 `MeStatsResponse` 의 필드를 long 으로 변경.
`Math.toIntExact` 변환 제거 → 미래 오버플로우 리스크 0.

### I5. 컨트롤러 메서드 네이밍
`getMyStats` → `myStats` (Spring 컨벤션상 동사 prefix 회피).

### I6. DTO 직렬화 회귀 테스트
`MeStatsResponseTest` 3건 — 모든 필드 + 중첩 WeekRange + LocalDate ISO 직렬화,
null topGrade 유지, WeekRange 단독 직렬화.
Envelope 자동 래핑은 `GlobalResponseWrapperTest` 에서 이미 커버.

### I4 스킵
WeekRange 페이로드 포함은 MVP 유지 — 리뷰어도 "MVP 에선 OK" 명시.

### 검증
- `:crimp-common:test :crimp-domain:test :crimp-api:test` 전체 pass (신규 3건 포함)
- 기존 MeStatsServiceTest 5건도 Optional 시그니처에 맞춰 업데이트.

+++ api/crimp-domain/src/main/java/io/crimp/domain/user/MeStatsView.java
@@ -5,10 +5,12 @@ import java.time.LocalDate;
 /**
  * 홈 대시보드용 사용자 집계 뷰 DTO.
  * weekRange 는 컨트롤러 레이어에서 nested 객체로 변환한다.
+ *
+ * 모든 카운트는 long 으로 통일 — 먼 미래에도 오버플로우 걱정 없음.
  */
 public record MeStatsView(
-        int weekSessions,
-        int weekSends,
+        long weekSessions,
+        long weekSends,
         long totalSessions,
         long totalSends,
         String topGrade,
